Refactor out color helpers
