+++ test/specs/core.helpers.tests.js
@@ -6,172 +6,6 @@ describe('Core helper tests', function() {
 		helpers = window.Chart.helpers;
 	});
 
-	it('should merge a normal config without scales', function() {
-		var baseConfig = {
-			valueProp: 5,
-			arrayProp: [1, 2, 3, 4, 5, 6],
-			objectProp: {
-				prop1: 'abc',
-				prop2: 56
-			}
-		};
-
-		var toMerge = {
-			valueProp2: null,
-			arrayProp: ['a', 'c'],
-			objectProp: {
-				prop1: 'c',
-				prop3: 'prop3'
-			}
-		};
-
-		var merged = helpers.configMerge(baseConfig, toMerge);
-		expect(merged).toEqual({
-			valueProp: 5,
-			valueProp2: null,
-			arrayProp: ['a', 'c'],
-			objectProp: {
-				prop1: 'c',
-				prop2: 56,
-				prop3: 'prop3'
-			}
-		});
-	});
-
-	it('should merge scale configs', function() {
-		var baseConfig = {
-			scales: {
-				prop1: {
-					abc: 123,
-					def: '456'
-				},
-				prop2: 777,
-				yAxes: [{
-					type: 'linear',
-				}, {
-					type: 'log'
-				}]
-			}
-		};
-
-		var toMerge = {
-			scales: {
-				prop1: {
-					def: 'bbb',
-					ghi: 78
-				},
-				prop2: null,
-				yAxes: [{
-					type: 'linear',
-					axisProp: 456
-				}, {
-					// pulls in linear default config since axis type changes
-					type: 'linear',
-					position: 'right'
-				}, {
-					// Pulls in linear default config since axis not in base
-					type: 'linear'
-				}]
-			}
-		};
-
-		var merged = helpers.configMerge(baseConfig, toMerge);
-		expect(merged).toEqual({
-			scales: {
-				prop1: {
-					abc: 123,
-					def: 'bbb',
-					ghi: 78
-				},
-				prop2: null,
-				yAxes: [{
-					type: 'linear',
-					axisProp: 456
-				}, {
-					display: true,
-
-					gridLines: {
-						color: 'rgba(0, 0, 0, 0.1)',
-						drawBorder: true,
-						drawOnChartArea: true,
-						drawTicks: true, // draw ticks extending towards the label
-						tickMarkLength: 10,
-						lineWidth: 1,
-						offsetGridLines: false,
-						display: true,
-						zeroLineColor: 'rgba(0,0,0,0.25)',
-						zeroLineWidth: 1,
-						zeroLineBorderDash: [],
-						zeroLineBorderDashOffset: 0.0,
-						borderDash: [],
-						borderDashOffset: 0.0
-					},
-					position: 'right',
-					offset: false,
-					scaleLabel: Chart.defaults.scale.scaleLabel,
-					ticks: {
-						beginAtZero: false,
-						minRotation: 0,
-						maxRotation: 50,
-						mirror: false,
-						padding: 0,
-						reverse: false,
-						display: true,
-						callback: merged.scales.yAxes[1].ticks.callback, // make it nicer, then check explicitly below
-						autoSkip: true,
-						autoSkipPadding: 0,
-						labelOffset: 0,
-						minor: {},
-						major: {},
-					},
-					type: 'linear'
-				}, {
-					display: true,
-
-					gridLines: {
-						color: 'rgba(0, 0, 0, 0.1)',
-						drawBorder: true,
-						drawOnChartArea: true,
-						drawTicks: true, // draw ticks extending towards the label,
-						tickMarkLength: 10,
-						lineWidth: 1,
-						offsetGridLines: false,
-						display: true,
-						zeroLineColor: 'rgba(0,0,0,0.25)',
-						zeroLineWidth: 1,
-						zeroLineBorderDash: [],
-						zeroLineBorderDashOffset: 0.0,
-						borderDash: [],
-						borderDashOffset: 0.0
-					},
-					position: 'left',
-					offset: false,
-					scaleLabel: Chart.defaults.scale.scaleLabel,
-					ticks: {
-						beginAtZero: false,
-						minRotation: 0,
-						maxRotation: 50,
-						mirror: false,
-						padding: 0,
-						reverse: false,
-						display: true,
-						callback: merged.scales.yAxes[2].ticks.callback, // make it nicer, then check explicitly below
-						autoSkip: true,
-						autoSkipPadding: 0,
-						labelOffset: 0,
-						minor: {},
-						major: {},
-					},
-					type: 'linear'
-				}]
-			}
-		});
-
-		// Are these actually functions
-		expect(merged.scales.yAxes[1].ticks.callback).toEqual(jasmine.any(Function));
-		expect(merged.scales.yAxes[2].ticks.callback).toEqual(jasmine.any(Function));
-	});
-
 	it('should filter an array', function() {
 		var data = [-10, 0, 6, 0, 7];
 		var callback = function(item) {
@@ -810,55 +644,5 @@ describe('Core helper tests', function() {
 		expect(helpers.getMaximumWidth(canvas)).toBe(280);
 
 		document.body.removeChild(div);
-	});
-
-	describe('Color helper', function() {
-		function isColorInstance(obj) {
-			return typeof obj === 'object' && obj.hasOwnProperty('values') && obj.values.hasOwnProperty('rgb');
-		}
-
-		it('should return a color when called with a color', function() {
-			expect(isColorInstance(helpers.color('rgb(1, 2, 3)'))).toBe(true);
-		});
-
-		it('should return a color when called with a CanvasGradient instance', function() {
-			var context = document.createElement('canvas').getContext('2d');
-			var gradient = context.createLinearGradient(0, 1, 2, 3);
-
-			expect(isColorInstance(helpers.color(gradient))).toBe(true);
-		});
-	});
-
-	describe('Background hover color helper', function() {
-		it('should return a CanvasPattern when called with a CanvasPattern', function(done) {
-			var dots = new Image();
-			dots.src = 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAA4AAAAOCAMAAAAolt3jAAAAD1BMVEUAAAD///////////////+PQt5oAAAABXRSTlMAHlFhZsfk/BEAAAAqSURBVHgBY2BgZGJmYmSAAUYWEIDzmcBcJhiXGcxlRpPFrhdmMiqgvX0AcGIBEUAo6UAAAAAASUVORK5CYII=';
-			dots.onload = function() {
-				var chartContext = document.createElement('canvas').getContext('2d');
-				var patternCanvas = document.createElement('canvas');
-				var patternContext = patternCanvas.getContext('2d');
-				var pattern = patternContext.createPattern(dots, 'repeat');
-				patternContext.fillStyle = pattern;
-
-				var backgroundColor = helpers.getHoverColor(chartContext.createPattern(patternCanvas, 'repeat'));
-
-				expect(backgroundColor instanceof CanvasPattern).toBe(true);
-
-				done();
-			};
-		});
-
-		it('should return a CanvasGradient when called with a CanvasGradient', function() {
-			var context = document.createElement('canvas').getContext('2d');
-			var gradient = context.createLinearGradient(0, 1, 2, 3);
-
-			expect(helpers.getHoverColor(gradient) instanceof CanvasGradient).toBe(true);
-		});
-
-		it('should return a modified version of color when called with a color', function() {
-			var originalColorRGB = 'rgb(70, 191, 189)';
-
-			expect(helpers.getHoverColor('#46BFBD')).not.toEqual(originalColorRGB);
-		});
 	});
 });
